+++ common/src/main/java/net/mehvahdjukaar/polytone/colormap/IColorGetter.java
@@ -14,11 +14,11 @@ import org.jetbrains.annotations.Nullable;
 
 public interface IColorGetter extends BlockColor, BarColor {
 
-    default boolean needsToFillTexture(){
+    default boolean needsToFillTexture() {
         return false;
     }
 
-    default IColorGetter makeConcurrent(){
+    default IColorGetter makeConcurrent() {
         return this;
     }
 
@@ -37,15 +37,24 @@ public interface IColorGetter extends BlockColor, BarColor {
             BlockState state = world.getBlockState(pos);
             return bc.getColor(state, world, pos, i) | 0xff000000;
         }
+
+        @Override
+        public int sampleColor(@Nullable BlockState state, @Nullable BlockPos pos, @Nullable Biome biome, @Nullable ItemStack item) {
+            if (state != null && pos != null) {
+                return bc.getColor(state, null, pos, 0) | 0xff000000;
+            }
+            return -1;
+        }
     }
 
     //wraps around a color resolver. note that usually the block color get color internally calls the color resolver itself which with grass replacement might be us
-    record ofColorResolver(BlockColor bc, ColorResolver cr) implements  IColorGetter,ColorResolver{
+    record ofColorResolver(BlockColor bc, ColorResolver cr) implements IColorGetter, ColorResolver {
 
         @Override
         public int getColor(BlockState state, @Nullable BlockAndTintGetter reader, @Nullable BlockPos pos, int tintIndex) {
             return bc.getColor(state, reader, pos, tintIndex);
         }
+
         @Override
         public int getItemColor(ItemStack stack, int tintIndex) {
             Minecraft mc = Minecraft.getInstance();
@@ -58,10 +67,19 @@ public interface IColorGetter extends BlockColor, BarColor {
         }
 
         @Override
-        public int getColor(Biome biome, double d, double e) {
-            return cr.getColor(biome, d, e);
+        public int getColor(Biome biome, double x, double z) {
+            return cr.getColor(biome, x, z);
         }
 
+        @Override
+        public int sampleColor(@Nullable BlockState state, @Nullable BlockPos pos, @Nullable Biome biome, @Nullable ItemStack item) {
+            if (biome != null) {
+                int x = pos == null ? 0 : pos.getX();
+                int z = pos == null ? 0 : pos.getZ();
+                return cr.getColor(biome, x, z);
+            }
+            return -1;
+        }
     }
 
     record OfItem(BarColor ic) implements IColorGetter {
@@ -74,5 +92,13 @@ public interface IColorGetter extends BlockColor, BarColor {
         public int getItemColor(ItemStack itemStack, int i) {
             return ic.getItemColor(itemStack, i);
         }
+
+        @Override
+        public int sampleColor(@Nullable BlockState state, @Nullable BlockPos pos, @Nullable Biome biome, @Nullable ItemStack item) {
+            return ic.getItemColor(item == null ? ItemStack.EMPTY : item, 0);
+        }
     }
+
+
+    int sampleColor(@Nullable BlockState state, @Nullable BlockPos pos, @Nullable Biome biome, @Nullable ItemStack item);
 }